Refactor process pool
