+++ clowder/clowder/git/repo.py
@@ -11,9 +11,8 @@ from termcolor import colored, cprint
 
 import clowder.util.formatting as fmt
 from clowder.error.clowder_git_error import ClowderGitError
-from clowder.util.execute import execute_command
+from clowder.process_pool import execute_command
 from clowder.util.file_system import remove_directory
-
 
 DEFAULT_REF = 'refs/heads/master'
 DEFAULT_REMOTE = 'origin'
