+++ clowder/clowder/source.py
