+++ test/unittests/test_fork.py
@@ -4,8 +4,8 @@ import os
 import sys
 import unittest
 
-from clowder.model.fork import Fork
-from clowder.model.source import Source
+from clowder.fork import Fork
+from clowder.source import Source
 from unittests.shared import GITHUB_SSH_SOURCE_YAML
