+++ test/unittests/test_project.py
@@ -4,8 +4,8 @@ import os
 import sys
 import unittest
 
-from clowder.model.project import Project
-from clowder.model.source import Source
+from clowder.project import Project
+from clowder.source import Source
 from unittests.shared import (
     DEFAULTS_YAML,
     GITHUB_HTTPS_SOURCE_YAML,
