+++ test/unittests/test_source.py
@@ -3,7 +3,7 @@
 import sys
 import unittest
 
-from clowder.model.source import Source
+from clowder.source import Source
 from unittests.shared import (
     GITHUB_HTTPS_SOURCE_YAML,
     GITHUB_SSH_SOURCE_YAML
